Add file structure
