+++ app/src/main/java/com/scouts/kitchenplaner/ui/view/inviteToProject/InviteToProject.kt
@@ -14,16 +14,17 @@
  * GNU General Public License for more details.
  */
 
-package com.scouts.kitchenplaner.ui.view
+package com.scouts.kitchenplaner.ui.view.inviteToProject
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun RecipeOverview(){
+fun InviteToProject(){
     Column {
-        Text(text = "This is the recipe overview, where all recipes are displayed")
+        Text("After project creation people can be invited")
         Text(text= "available Links to other sides are: ")
+
     }
 }
